@@ -1,11 +1,6 @@
-import Vue from 'vue'
+import { createApp } from 'vue'
+import './styles/base_layout.css'
+import './styles/welcome.css'
 import App from './App.vue'
-import vuetify from './plugins/vuetify'
-import './ml'
 
-Vue.config.productionTip = false
-
-new Vue({
-  vuetify,
-  render: h => h(App)
-}).$mount('#app')
+createApp(App).mount('#app')
